pyproject.toml
[tool.poetry]
name = "ctf-k8"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
cloudflare = "^2.8.15"
requests = "^2.25.1"
boto3 = "^1.17.79"
docker = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
